refactor: attempt to generate CSVs after k8s manifests are created
This is a sync commit.

The idea is that we need to be able to read any existing roles or
deployments to add that information to the generated CSV. However, so
far, either the CSV generation occurs before the manifests are
generated, doesn't occur at all or introduces a build item cycle (which
is the case with this specific version of the code).