feat: support wider shared csv metadata
Renamed CSVGroupName to CSVMetadata and added more fields. Introduced
the SharedCSVMetadata interface that controllers wishing to share
metadata, and therefore be part of the same generated CSV, need to
implement, parameterized by an object that needs to be annotated with
@CSVMetadata. It's still possible to annotate controllers with
CSVMetadata directly or even override values from SharedCSVMetadata by
individual values found in the direct CSVMetadata annotation. The
name field from CSVMetadata controls the name of the generated CSV so
all controllers using the same CSVMetadata.name will be part of the same
generated CSV file.