feat: add adequate Metrics instance to ConfigurationService
We register a `MeterBinder` implementation that gets called
automatically by the Quarkus micrometer extension if a
micrometer-compatible extension is provided by the operator app. That
binder implementation simply creates a new Metrics instance with the
provided `MeterRegistry` implementation. We then inject that Metrics
instance into our QuarkusConfigurationService.

Thanks to @ebullient for the missing insight!