feat: add CSV Generation extension
Note that this relies on a trick to make sure the extension executes: as
it doesn't contribute anything to the build directly, Quarkus would
actually skip it. The trick is to make the final step produce a feature
so that it would always be triggered. That feels a little dirty, though.