fix: start the operator when we register the delayed controller
This is needed because prior to the 1.9.8 upgrade, registering a
controller would automatically start it, thus calling its init method,
allowing us to set its initialized field for our tests. As this isn't
possible anymore, work around this issue by starting the operator, which
will start the controllers once they are registered, thus setting the
initialized field as expected.