fix: RBAC generation (#150)
* fix: generate proper RBACs depending on config

We generate a RoleBinding without namespace information when the
operator is watching the namespace it's deployed in with the assumption
that the manifest will be applied to the appropriate namespace. When the
operator watches all namespaces, we generate a ClusterRoleBinding but
we cannot provide the required namespace to identify the ServiceAccount
used by the operator as it is not known at build time. When the operator
watches several namespaces, we generate a RoleBinding for each namespace

Also add permissions bindings to validate CRDs if the option is activated.