Fix RBAC generation issues (#172)
* feat: also process namespaces at build time

This is needed because RBAC generation needs to make decisions based on
which namespaces the controllers are configured to watch.

* feat: make it possible to disable RBAC generation altogether

* fix: handle unexpanded namespaces property

The property was removed from BuildTimeControllerConfiguration because
it present on a `@ConfigRoot` would trigger the expansion of variables.
This new implementation sidesteps this issue and currently generates
the RBAC with the variable name e.g. `${foo}`.

* fix: handle quoted controller name in application.properties

* fix: generate CustomResourceInfos if CRD generation is disabled

* fix: generate ClusterRoleBinding for CRD permissions, warning output

Also take `quarkus.kubernetes.namespace` property into account.