feat!: explicit interaction between CRD generate/apply and launch mode
It is now impossible to automatically apply the CRD to the cluster in
normal mode, even if explicitly asked for, to prevent potential issues
in production. CRD is automatically applies in dev or test mode now,
though, which is different from the previous behavior, where you needed
to activate the feature explicitly.

Fixes #297