feat: look up icons in src/main/kubernetes, do not use base64 data
A warning is issued if we detect legacy usage of default icon in the
app's resources. Icons should now be put in `src/main/kubernetes` to
prevent them from being included in the app's binary. We still support
using a default icon that will be picked up automatically if named using
the `<reconciler name>-operator.icon.png` name.