fix: make sure dependent beans are registered in dev mode
It could happen that, during iterative development, a controller's
configuration might not be completely resolved, in particular when it
comes to the dependent resources. However, on subsequent live reload,
the configuration is found and, if no classes impacted the generation of
that specific configuration, dependents would not be processed again,
resulting in the beans not being known to Arc when needed.