fix: use deployment namespace if provided when generating RoleBindings (#456)
We need to make sure we run after `AddNamespaceDecorator` because
that decorator will overwrite any namespaces we might have set on our
`RoleBindings` if we let it run after our own decorator. 

Proper behavior for the `AddNamespaceDecorator` should be to not 
overwrite namespaces if they have been set. This has been fixed in
https://github.com/quarkusio/quarkus/pull/29813