Enable OperatorSDKTest
The reason it was previously failing is that
Dekorate could not find a valid Maven project
in the temp directory in which QuarkusProdModeTest
bootstraps the Quarkus application.
In order to make it work, we need to ensure that
the aforementioned temp directory is created under
the project root, so Dekorate then finds it by moving
up the file system.

The same thing is done in the Quarkus repo to
make the Kubernetes related QuarkusProdModeTest
tests work.
The reason argLine is used instead of systemProperties
is mentioned in the comment