fix: Treat empty runtime namespaces as All Namespaces mode
If the runtime configuration specifies an empty string as the namespaces to watch, we should treat that as all namespaces mode, mainly because that's how OLM signifies that an operator should watch all namespaces.

An empty string is treated the same as unset by the configuration parser, so we need to set a default on the field so we can explicitly check for it being unset, to determine if the empty string was used.

Fixes #656

Signed-off-by: James Hewitt <james.hewitt@uk.ibm.com>