fix: helm chart generates invalid YAMl when watching all namespaces
When a reconciler is configured to watch all namespaces,
a ClusterRoleBinding object is generated by
AddRoleBindingsDecorator, and it later gets appended
to templates/kubernetes.yml by HelmChartProcessor.

However, that file becomes syntactically invalid because
FileUtils.asYaml produces a top-level YAML array when passed
a List. That's not what Helm expects.

When one tries to install the generated chart, the following
error is returned:

   Error: INSTALLATION FAILED: YAML parse error on
   [...]/templates/kubernetes.yml: error unmarshaling JSON:
   while decoding JSON: json: cannot unmarshal array into
   Go value of type releaseutil.SimpleHead
   YAML parse error on [...]/templates/kubernetes.yml