feat: separate build time controller configuration from runtime
Previously, the controller configuration used to carry over all the
information that was needed for build time processing and required
reconfiguration at runtime of AnnotationConfigurables. We removed the
unneeded information, which should slightly decrease memory usage and
avoid exposing useless information. AnnotationConfigurables are now
processed at build time. The downside of this is that implementations
now need to be byte-recordable for configuration to properly carry over
to runtime.